+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 # JSON to Schema Changelog
 
-## [Unreleased]
+## [1.1]
+- Support for 2021.3 platform version
+- Support for 'null' type
+- The action is now available in indexing mode
+- Auto save before schema generation